feat:admit parent issue as a query param
