+++ .prettierrc
@@ -1,6 +1,6 @@
 {
   "trailingComma": "es5",
   "tabWidth": 2,
-  "semi": false,
+  "semi": true,
   "singleQuote": false
 }
