+++ src/main.ts
@@ -2,7 +2,7 @@ import { main } from "./core"
 import { Agent, setGlobalDispatcher } from "undici"
 import { Logger } from "./model/logger"
 import { Response } from "./model/response"
-
+import { Request } from "express"
 // Ignore self signed certs
 const agent = new Agent({
   connect: {
@@ -16,7 +16,7 @@ const express = require("express")
 const app = express()
 const port = 3000
 
-app.get("/:id", async (req: any, res: any) => {
+app.get("/:id", async (req: Request, res: any) => {
   const { id } = req.params
   if (isNaN(+id) || +id <= 0) {
     return
@@ -29,7 +29,7 @@ app.get("/:id", async (req: any, res: any) => {
   }
 
   try {
-    const result = await main(+id, logger)
+    const result = await main(+id, req.query, logger)
     res.send(result)
   } catch (e) {
     logger.error("General error")
